+++ R/gg-operators.R
@@ -61,8 +61,10 @@ print.arphit.gg <- function(x, ...) {
          "yaxislabel" = addaxislabel(gg, element, "y"),
          "xaxislabel" = addaxislabel(gg, element, "x"),
          "legend" = addlegend(gg, element),
-         "autolabel" = enableautolabel(gg, element$quiet, element$arrow_lines, element$arrow_bars),
+         "autolabel" = enableautolabel(gg, element$quiet, element$arrow_lines,
+                                       element$arrow_bars, element$ignore_existing_labels),
          "xfreq" = addxfreq(gg, element$freq, element$panel),
-         stop("Unknown element type for arphit.gg", call. = FALSE))
+         "rename_series" = renameseries(gg, element$mapping, element$panel),
+         stop("Unknown element type for arphitgg", call. = FALSE))
   return(gg)
 }